Redux forms now working, creater main component for Todo and Fridge.
