--- src/components/TodoComplete.js
@@ -1,20 +1,0 @@
-import React, { Component } from 'react'
-
-class TodoComplete extends Component {
-    render () {
-        const { onCompleteClick } = this.props
-
-        return (
-            <div className='container'>
-                <button
-                    onClick={() => onCompleteClick()}
-                    className='btn btn-danger'
-                >
-                    Complete todo
-                </button>
-            </div>
-        )
-    }
-}
-
-export default TodoComplete
